# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Mevi Commerce Hub, an agricultural commerce and order booking platform.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductStatus = typeof ProductStatus[keyof typeof ProductStatus];


export const ProductStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Product {
  id: number;
  name: string;
  sku: string;
  category: string;
  description: string;
  origin: string;
  packageSize: string;
  shelfLife: string;
  storageInstructions: string;
  imageUrls: string[];
  /** @minimum 0 */
  price: number;
  /**
     * @minimum 0
     * @nullable
     */
  salePrice?: number | null;
  status: ProductStatus;
  /** @nullable */
  rejectionReason?: string | null;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  sku: string;
  /** @minLength 1 */
  category: string;
  description: string;
  origin: string;
  packageSize: string;
  shelfLife: string;
  storageInstructions: string;
  imageUrls: string[];
  /** @minimum 0 */
  price: number;
  /**
     * @minimum 0
     * @nullable
     */
  salePrice?: number | null;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  sku?: string;
  /** @minLength 1 */
  category?: string;
  description?: string;
  origin?: string;
  packageSize?: string;
  shelfLife?: string;
  storageInstructions?: string;
  imageUrls?: string[];
  /** @minimum 0 */
  price?: number;
  /**
     * @minimum 0
     * @nullable
     */
  salePrice?: number | null;
}

export type ProductModerationInputStatus = typeof ProductModerationInputStatus[keyof typeof ProductModerationInputStatus];


export const ProductModerationInputStatus = {
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface ProductModerationInput {
  status: ProductModerationInputStatus;
  /** @nullable */
  rejectionReason?: string | null;
}

export interface AgencyParams {
  /** @minLength 1 */
  slug: string;
}

export interface OrderItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 6 */
  phone: string;
  /** @minLength 1 */
  address: string;
  /** @minLength 1 */
  province: string;
  /** @minLength 1 */
  district: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  agencySlug?: string | null;
  /** @minItems 1 */
  items: OrderItemInput[];
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  lineTotal: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  new: 'new',
  confirmed: 'confirmed',
  processing: 'processing',
  shipped: 'shipped',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  code: string;
  customerName: string;
  phone: string;
  address: string;
  province: string;
  district: string;
  /** @nullable */
  note: string | null;
  total: number;
  status: OrderStatus;
  /** @nullable */
  agencySlug: string | null;
  items: OrderItem[];
  createdAt: string;
}

export type OrderStatusInputStatus = typeof OrderStatusInputStatus[keyof typeof OrderStatusInputStatus];


export const OrderStatusInputStatus = {
  new: 'new',
  confirmed: 'confirmed',
  processing: 'processing',
  shipped: 'shipped',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusInput {
  status: OrderStatusInputStatus;
}

export interface AdminSummary {
  productCount: number;
  pendingProducts: number;
  orderCount: number;
  newOrders: number;
  totalRevenue: number;
  recentOrders: Order[];
}

export interface Agency {
  id: number;
  slug: string;
  displayName: string;
  /** @nullable */
  logoUrl: string | null;
  primaryColor: string;
  contactName: string;
  contactPhone: string;
  contactEmail: string;
  active: boolean;
}

export interface PublicAgency {
  id: number;
  slug: string;
  displayName: string;
  /** @nullable */
  logoUrl: string | null;
  primaryColor: string;
  active: boolean;
}

export interface AgencyInput {
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  displayName: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @pattern ^#[0-9A-Fa-f]{6}$ */
  primaryColor: string;
  contactName: string;
  contactPhone: string;
  contactEmail: string;
}

export interface AgencyUpdate {
  /** @minLength 1 */
  slug?: string;
  /** @minLength 1 */
  displayName?: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @pattern ^#[0-9A-Fa-f]{6}$ */
  primaryColor?: string;
  contactName?: string;
  contactPhone?: string;
  contactEmail?: string;
  active?: boolean;
}

export interface AgencySummary {
  orderCount: number;
  openOrders: number;
  totalRevenue: number;
  recentOrders: Order[];
}

export type UploadRequestContentType = typeof UploadRequestContentType[keyof typeof UploadRequestContentType];


export const UploadRequestContentType = {
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'image/webp': 'image/webp',
} as const;

export interface UploadRequest {
  name: string;
  /** @minimum 1 */
  size: number;
  contentType: UploadRequestContentType;
}

export interface UploadResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListProductsParams = {
q?: string;
category?: string;
/**
 * @minimum 0
 */
minPrice?: number;
/**
 * @minimum 0
 */
maxPrice?: number;
sort?: ListProductsSort;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  featured: 'featured',
  newest: 'newest',
  'price-asc': 'price-asc',
  'price-desc': 'price-desc',
} as const;

export type ListAdminProductsParams = {
status?: ListAdminProductsStatus;
};

export type ListAdminProductsStatus = typeof ListAdminProductsStatus[keyof typeof ListAdminProductsStatus];


export const ListAdminProductsStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  all: 'all',
} as const;

export type ListAdminOrdersParams = {
status?: ListAdminOrdersStatus;
};

export type ListAdminOrdersStatus = typeof ListAdminOrdersStatus[keyof typeof ListAdminOrdersStatus];


export const ListAdminOrdersStatus = {
  new: 'new',
  confirmed: 'confirmed',
  processing: 'processing',
  shipped: 'shipped',
  completed: 'completed',
  cancelled: 'cancelled',
  all: 'all',
} as const;