# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for Mevi Commerce Hub, an agricultural commerce and order booking platform.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/summary:
    get:
      operationId: getAdminSummary
      tags: [admin]
      summary: Get administrator dashboard metrics
      responses:
        "200":
          description: Overview of catalog and order activity
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminSummary" }
        "401": { description: Sign in required }
        "403": { description: Administrator role required }
  /admin/products:
    get:
      operationId: listAdminProducts
      tags: [admin]
      summary: List products for moderation and management
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected, all]
      responses:
        "200":
          description: Products managed by Mevi
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
    post:
      operationId: createProduct
      tags: [admin]
      summary: Create a catalog product
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductInput" }
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
  /admin/products/{id}:
    patch:
      operationId: updateProduct
      tags: [admin]
      summary: Update a catalog product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductUpdate" }
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404": { description: Product not found }
    delete:
      operationId: deleteProduct
      tags: [admin]
      summary: Delete a catalog product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Product deleted }
        "404": { description: Product not found }
  /admin/products/{id}/approval:
    post:
      operationId: moderateProduct
      tags: [admin]
      summary: Approve or reject a submitted product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductModerationInput" }
      responses:
        "200":
          description: Moderation result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404": { description: Product not found }
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List all order bookings
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [new, confirmed, processing, shipped, completed, cancelled, all]
      responses:
        "200":
          description: Orders across all channels
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
  /admin/orders/{id}:
    patch:
      operationId: updateOrderStatus
      tags: [admin]
      summary: Update order processing status
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderStatusInput" }
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "404": { description: Order not found }
  /admin/agencies:
    get:
      operationId: listAgencies
      tags: [admin]
      summary: List agency stores
      responses:
        "200":
          description: Agency directory
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Agency" }
    post:
      operationId: createAgency
      tags: [admin]
      summary: Create a white-label agency store
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgencyInput" }
      responses:
        "201":
          description: Agency created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agency" }
  /admin/agencies/{id}:
    patch:
      operationId: updateAgency
      tags: [admin]
      summary: Update an agency store
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgencyUpdate" }
      responses:
        "200":
          description: Updated agency
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agency" }
        "404": { description: Agency not found }
  /agency/summary:
    get:
      operationId: getAgencySummary
      tags: [agency]
      summary: Get the current agency's order metrics
      responses:
        "200":
          description: Agency dashboard overview
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgencySummary" }
        "401": { description: Sign in required }
        "403": { description: Agency role required }
  /agency/settings:
    get:
      operationId: getAgencySettings
      tags: [agency]
      summary: Get the current agency's branding
      responses:
        "200":
          description: Agency profile and brand
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agency" }
    patch:
      operationId: updateMyAgency
      tags: [agency]
      summary: Update the current agency's branding and contact details
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgencyUpdate" }
      responses:
        "200":
          description: Updated agency profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agency" }
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, sku, category, description, origin, packageSize, shelfLife, storageInstructions, imageUrls, price, status, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        sku: { type: string }
        category: { type: string }
        description: { type: string }
        origin: { type: string }
        packageSize: { type: string }
        shelfLife: { type: string }
        storageInstructions: { type: string }
        imageUrls:
          type: array
          items: { type: string }
        price: { type: integer, minimum: 0 }
        salePrice: { type: ["integer", "null"], minimum: 0 }
        status: { type: string, enum: [pending, approved, rejected] }
        rejectionReason: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    ProductInput:
      type: object
      required: [name, sku, category, description, origin, packageSize, shelfLife, storageInstructions, imageUrls, price]
      properties:
        name: { type: string, minLength: 1 }
        sku: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        description: { type: string }
        origin: { type: string }
        packageSize: { type: string }
        shelfLife: { type: string }
        storageInstructions: { type: string }
        imageUrls:
          type: array
          items: { type: string }
        price: { type: integer, minimum: 0 }
        salePrice: { type: ["integer", "null"], minimum: 0 }
    ProductUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        sku: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        description: { type: string }
        origin: { type: string }
        packageSize: { type: string }
        shelfLife: { type: string }
        storageInstructions: { type: string }
        imageUrls:
          type: array
          items: { type: string }
        price: { type: integer, minimum: 0 }
        salePrice: { type: ["integer", "null"], minimum: 0 }
    ProductModerationInput:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [approved, rejected] }
        rejectionReason: { type: ["string", "null"] }
...
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, unitPrice, lineTotal]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: integer }
        unitPrice: { type: integer }
        lineTotal: { type: integer }
    Order:
      type: object
      required: [id, code, customerName, phone, address, province, district, note, total, status, agencySlug, items, createdAt]
      properties:
        id: { type: integer }
        code: { type: string }
        customerName: { type: string }
        phone: { type: string }
        address: { type: string }
        province: { type: string }
        district: { type: string }
        note: { type: ["string", "null"] }
        total: { type: integer }
        status: { type: string, enum: [new, confirmed, processing, shipped, completed, cancelled] }
        agencySlug: { type: ["string", "null"] }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        createdAt: { type: string, format: date-time }
    OrderStatusInput:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [new, confirmed, processing, shipped, completed, cancelled] }
    AdminSummary:
      type: object
      required: [productCount, pendingProducts, orderCount, newOrders, totalRevenue, recentOrders]
      properties:
        productCount: { type: integer }
        pendingProducts: { type: integer }
        orderCount: { type: integer }
        newOrders: { type: integer }
        totalRevenue: { type: integer }
        recentOrders:
          type: array
          items: { $ref: "#/components/schemas/Order" }
    Agency:
      type: object
      required: [id, slug, displayName, logoUrl, primaryColor, contactName, contactPhone, contactEmail, active]
      properties:
        id: { type: integer }
        slug: { type: string }
        displayName: { type: string }
        logoUrl: { type: ["string", "null"] }
        primaryColor: { type: string }
        contactName: { type: string }
        contactPhone: { type: string }
        contactEmail: { type: string }
        active: { type: boolean }
...
    AgencyInput:
      type: object
      required: [slug, displayName, primaryColor, contactName, contactPhone, contactEmail]
      properties:
        slug: { type: string, minLength: 1 }
        displayName: { type: string, minLength: 1 }
        logoUrl: { type: ["string", "null"] }
        primaryColor: { type: string, pattern: "^#[0-9A-Fa-f]{6}$" }
        contactName: { type: string }
        contactPhone: { type: string }
        contactEmail: { type: string }
    AgencyUpdate:
      type: object
      properties:
        slug: { type: string, minLength: 1 }
        displayName: { type: string, minLength: 1 }
        logoUrl: { type: ["string", "null"] }
        primaryColor: { type: string, pattern: "^#[0-9A-Fa-f]{6}$" }
        contactName: { type: string }
        contactPhone: { type: string }
        contactEmail: { type: string }
        active: { type: boolean }
    AgencySummary:
      type: object
      required: [orderCount, openOrders, totalRevenue, recentOrders]
      properties:
        orderCount: { type: integer }
        openOrders: { type: integer }
        totalRevenue: { type: integer }
        recentOrders:
          type: array
          items: { $ref: "#/components/schemas/Order" }
...