# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Mevi Commerce Hub, an agricultural commerce and order booking platform.
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  code: string;
  customerName: string;
  phone: string;
  address: string;
  province: string;
  district: string;
  /** @nullable */
  note: string | null;
  total: number;
  status: OrderStatus;
  /** @nullable */
  agencySlug: string | null;
  items: OrderItem[];
  createdAt: Date;
}