# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Mevi Commerce Hub, an agricultural commerce and order booking platform.
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  lineTotal: number;
}