# lib/api-zod/src/generated/types/orderStatusInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Mevi Commerce Hub, an agricultural commerce and order booking platform.
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatusInputStatus } from './orderStatusInputStatus';

export interface OrderStatusInput {
  status: OrderStatusInputStatus;
}