# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Mevi Commerce Hub, an agricultural commerce and order booking platform.
 * OpenAPI spec version: 0.1.0
 */
import type { ProductStatus } from './productStatus';

export interface Product {
  id: number;
  name: string;
  sku: string;
  category: string;
  description: string;
  origin: string;
  packageSize: string;
  shelfLife: string;
  storageInstructions: string;
  imageUrls: string[];
  /** @minimum 0 */
  price: number;
  /**
     * @minimum 0
     * @nullable
     */
  salePrice?: number | null;
  status: ProductStatus;
  /** @nullable */
  rejectionReason?: string | null;
  createdAt: Date;
}