# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Mevi Commerce Hub, an agricultural commerce and order booking platform.
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  sku: string;
  /** @minLength 1 */
  category: string;
  description: string;
  origin: string;
  packageSize: string;
  shelfLife: string;
  storageInstructions: string;
  imageUrls: string[];
  /** @minimum 0 */
  price: number;
  /**
     * @minimum 0
     * @nullable
     */
  salePrice?: number | null;
}