# lib/api-zod/src/generated/types/productModerationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Mevi Commerce Hub, an agricultural commerce and order booking platform.
 * OpenAPI spec version: 0.1.0
 */
import type { ProductModerationInputStatus } from './productModerationInputStatus';

export interface ProductModerationInput {
  status: ProductModerationInputStatus;
  /** @nullable */
  rejectionReason?: string | null;
}